### 2024-10-23 #7328 @crvernon
:wave: @buddyroo30 - while I am finding you an editor, please reduce the word count of your paper to around 1000 words.  Also, please add an open source license to your repository (see https://opensource.org/licenses).
