### 2024-10-24 #7328 @logological
@crvernon This submission presents a problem similar to the one we had with PYLIWC (#7119) in that it is targeted at users of a proprietary tool, [CanvasXpress](https://canvasxpress.org/).  (Note that CanvasXpress falsely claims to be [open source](https://opensource.org/osd); it is actually distributed under [a licence that covers only uses that are "personal and educational" and places heavy-handed restrictions on redistribution](https://canvasxpress.org/docs/license.html).)  I am concerned about our ability to find reviewers who can (legally) use CanvasXpress to test the software.  Before PYLIWC was rejected (for unrelated reasons), I tried and failed to find reviewers who held a licence to the proprietary software it depended on.

In any case, even if it's not strictly outside JOSS's scope, I'm not particularly comfortable editing a submission on a free/open-source tool whose only purpose is to facilitate the use of proprietary software.

What are your thoughts?
