### 2024-10-25 #7328 @buddyroo30
Hi,
Thank you for the feedback on our submission. I went ahead and added an open source license as requested and I will work to reduce the word count. As for the issue around the license for CanvasXpress itself, let me discuss that with my co-author and get back to you soon.
Best,Andrew Smith

Yahoo Mail: Search, Organize, Conquer

  On Thu, Oct 24, 2024 at 4:39 PM, Chris ***@***.***> wrote:

Thank you for raising this issue @logological. I am going to place this on in review with our larger editorial board as well. So let's hold off on finding reviewers.

@buddyroo30 - please feel free to provide any comment here to address the above.
