### 2024-10-28 #7328 @buddyroo30
Hi,
So I discussed the license issue around CanvasXpress itself with my collaborator and he feels its current license is reasonable and does not want to change it. First, practically, the vast majority of users of CanvasXpress are academic or not for profit users, so they would be able to freely use CanvasXpress (and the associated AI enhancements we discuss in our paper) without limitation. Second, the license for CanvasXpress itself is essentially what most of the other large visualization packages use (e.g. Plotly) so CanvasXpress's license is not anything out of the ordinary. Please let us know your thoughts and if you think you could proceed with the review of the paper.
thanks,Andrew Smith
    On Friday, October 25, 2024 at 09:57:54 AM EDT, Andrew Smith ***@***.***> wrote:

 Hi,
Thank you for the feedback on our submission. I went ahead and added an open source license as requested and I will work to reduce the word count. As for the issue around the license for CanvasXpress itself, let me discuss that with my co-author and get back to you soon.
Best,Andrew Smith

Yahoo Mail: Search, Organize, Conquer

  On Thu, Oct 24, 2024 at 4:39 PM, Chris ***@***.***> wrote:

Thank you for raising this issue @logological. I am going to place this on in review with our larger editorial board as well. So let's hold off on finding reviewers.

@buddyroo30 - please feel free to provide any comment here to address the above.
