### 2024-10-28 ac007c9
Added new service ask_generic that lets you just ask any prompt of the llm (no extra context is added like for ask)
