+++ app.py
@@ -289,6 +289,90 @@ def ask():
     else:
         return(retValsJson)
 
+@app.route('/ask_generic',methods=['GET','POST'])
+def ask_generic():
+
+    start = time.time()
+
+    version = request.values.get('version') # PROD, DEVTEST, etc.
+    prompt = request.values.get('prompt')
+    model = request.values.get('model')
+    max_new_tokens = request.values.get('max_new_tokens')
+    topp = request.values.get('topp')
+    temperature = request.values.get('temperature')
+    presence_penalty = request.values.get('presence_penalty')
+    frequency_penalty = request.values.get('frequency_penalty')
+    callback = request.values.get('callback')
+    curDatetime = str(datetime.datetime.now().strftime("%Y-%m-%d %H:%M"))
+
+    if utils.empty(prompt):
+        errResp = { 'text': "Error: you must provide a prompt for the LLM", 'success': False}
+        return(json.dumps(errResp))
+
+    if utils.empty(version):
+        version = 'UNKNOWN'
+
+    if utils.empty(model):
+        model = 'gpt-4-32k'
+
+    if utils.empty(max_new_tokens):
+        max_new_tokens = 1024
+
+    if utils.empty(topp):
+        topp = 1.0
+
+    if utils.empty(temperature):
+        temperature = 0.0
+
+    if utils.empty(presence_penalty):
+        presence_penalty = 0.0
+
+    if utils.empty(frequency_penalty):
+        frequency_penalty = 0.0
+
+    USE_MODEL = llm.get_model_type(model)
+    print("USE_MODEL = " + USE_MODEL + ", model = " + model)
+
+    resp = {}
+
+    generated_text = "NOTHINGYET"
+    if USE_MODEL == "titan":
+        generated_text = llm.generate_results_titan(prompt,model=model, max_token_count=int(max_new_tokens), topp=float(topp), temperature=float(temperature))
+    elif USE_MODEL == "mistral":
+        generated_text = llm.generate_results_mistral(prompt,model=model,max_token_count=int(max_new_tokens),topp=float(topp),temperature=float(temperature))
+    elif USE_MODEL == "llama31":
+        generated_text = llm.generate_results_llama31(prompt,model=model, max_token_count=int(max_new_tokens), topp=float(topp), temperature=float(temperature))
+    elif USE_MODEL == "anthropic":
+        generated_text = llm.generate_results_anthropic(prompt, model=model, max_token_count=int(max_new_tokens), topp=float(topp), temperature=float(temperature))
+    elif USE_MODEL == "openai":
+        generated_text = llm.generate_results_openai(prompt, model=model, max_new_tokens=int(max_new_tokens),topp=float(topp),temperature=float(temperature),presence_penalty=float(presence_penalty),frequency_penalty=float(frequency_penalty))
+    elif USE_MODEL == 'google_gemini':
+        generated_text = llm.generate_results_google_gemini(prompt,model=model,temperature=float(temperature))
+    elif USE_MODEL == 'ollama':
+        generated_text = llm.generate_results_ollama(prompt,model=model)
+    end = time.time()
+    llm_time_taken = end - start
+    if utils.empty(generated_text) or generated_text.strip() == "''":
+        errResp = { 'text': "Error: No response was generated by the LLM, please try again.", 'success': False }
+        return(json.dumps(errResp))
+    print("RAW generated_text:\n" + generated_text)
+
+    resp = { 'success': True,
+            'generated_text': generated_text,
+            'total_time_taken': llm_time_taken,
+            'prompt': prompt,
+            'datetime': curDatetime }
+
+    retValsJson = json.dumps(resp)
+
+    if not utils.empty(callback):
[…]
