### 2024-11-05 69341c0
Initial checkin of code for accuracy testing with Isaac's synthetic data that he generated using his co-pilot framework
