+++ synth_examples/exec_accuracy_tests.py
@@ -1,0 +1,255 @@
+from FlagEmbedding import BGEM3FlagModel
+import json
+import sys
+from pymilvus import MilvusClient
+from pymilvus.model.hybrid import BGEM3EmbeddingFunction
+from pymilvus import connections
+import numpy as np
+import math
+import statistics
+import os
+from openai import AzureOpenAI
+from dotenv import load_dotenv
+import tiktoken
+import utils
+
+load_dotenv()
+
+openai_enc = tiktoken.encoding_for_model('gpt-4o-global')
+
+def generate_results_openai(prompt, model='gpt-4o-global', max_new_tokens=1024, topp=1.0, temperature=0.1):
+
+    client = AzureOpenAI()
+
+    completion = client.chat.completions.create(
+        model=model,
+        max_tokens=max_new_tokens,
+        temperature=temperature,
+        top_p=topp,
+        messages=[
+         {
+            "role": "user",
+            "content": prompt,
+            },
+        ],
+    )
+    generated_text = completion.choices[0].message.content
+
+    return(generated_text)
+
+bge_m3_ef = BGEM3EmbeddingFunction(
+    model_name='BAAI/bge-m3', # Specify the model name
+    device='cpu', # Specify the device to use, e.g., 'cpu' or 'cuda:0'
+    use_fp16=False # Specify whether to use fp16. Set to `False` if `device` is `cpu`.
+)
+
+def getFewShots(milvusClient, prompt,numFewShots=25,filterPrompt=False,format='text'):
+
+    in_numFewShots = numFewShots
+    if filterPrompt:
+        numFewShots = numFewShots + 1
+
+    prompt = prompt.replace("\n", " ")
+
+    queries = [prompt]
+
+    query_embeddings = bge_m3_ef.encode_queries(queries)
+    #collection_name="few_shot_examples",
+    #collection_name="canvasxpress_docs",
+
+    res = milvusClient.search(
+        collection_name="few_shot_examples",
+        data=[query_embeddings["dense"][0]],
+        limit=numFewShots,
+        output_fields=["config", "configEnglish", "headers", "id"],
+    )
+    #English Text:...; Headers/Column Names: ..., Answer: ...
+    #See here for where got below to get the hits: https://github.com/milvus-io/pymilvus/blob/master/examples/milvus_client/simple.py#L68
+    fewShotTxt = ""
+    fewShotObj = []
+    fewShotsCt = 0
+    for hits in res:
+        for hit in hits:
+            if fewShotsCt < in_numFewShots:
+                curConfig = hit['entity']['config'].replace("\n"," ")
+                curEnglishConfig = hit['entity']['configEnglish'].replace("\n", " ")
+                if filterPrompt and prompt == curEnglishConfig:
+                    continue
+                curHeadersColumnNames = hit['entity']['headers'].replace("\n", " ")
+                if format == 'text':
+                    fewShotTxt = fewShotTxt + f"English Text: {curEnglishConfig}; Headers/Column Names: {curHeadersColumnNames}, Answer: {curConfig}" + "\n"
+                elif format == 'json':
+                    curRec = { 'English Text': curEnglishConfig, 'Headers/Column Names': curHeadersColumnNames, 'Answer': curConfig }
+                    fewShotObj.append(curRec)
+                fewShotsCt = fewShotsCt + 1
+
+    if format == 'json':
+        fewShotTxt = json.dumps(fewShotObj)
+
+    return(fewShotTxt)
+
+def gen_vectordb(train_docs, num_docs, vectorDbFile):
+    """
[…]
