### 2024-11-05 40c38d0
Removing schema_info.txt, those will differ for different tests so don't save globally
