+++ synth_examples/schema_info.txt
@@ -1,84 +1,0 @@
-Here is detailed information about the CanvasXpress JSON configuration fields:
-graphType: Description: 'Specifies the type of graph', Type: 'string', Category: 'General', Options for Field Value: [Bar,Line,Area,AreaLine,BarLine,Boxplot,Dotplot,DotLine,Fish,Gantt,Meter,Heatmap,Stacked,StackedLine,StackedPercent,StackedPercentLine,Tree,Treemap,TagCloud,ParallelCoordinates,Sankey,Scatter2D,ScatterBubble2D,Scatter3D,Correlation,Pie,Venn,Network,Genome,Circular], Default Value: 'Bar'
-graphOrientation: Description: 'Specifies the orientation of one dimensional graphs', Type: 'string', Category: 'General', Options for Field Value: [horizontal,vertical], Default Value: 'horizontal'
-theme: Description: 'Themes for the plot. This parameter will overwrite colors, background, axes, etc', Type: 'string', Category: 'General', Options for Field Value: [none,grey,gray,igray,ggplot,bw,linedraw,light,dark,minimal,classic,void,economist,excel,tableau,stata,solarized,ptol,wsj,ggblanket,cx], Default Value: 'none'
-title: Description: 'Title of the graph', Type: 'string', Category: 'Titles and Subtitles', Default Value: 'False'
-titleColor: Description: 'Color for the text in title', Type: 'color', Category: 'Titles and Subtitles', Default Value: 'rgb(0,0,0)'
-titleScaleFontFactor: Description: 'Scaling factor used to increase or decrease the title font size in the canvas', Type: 'float', Category: 'Titles and Subtitles', Default Value: '1'
-colorScheme: Description: 'Color schemes can be user defined which will take the colors in the color property or one provided in canvasXpress. The order of the colors will be used to sequentially select when a different color is needed in a particular visualization', Type: 'string', Category: 'Colors', Options for Field Value: [YlGn,YlGnBu,GnBu,BuGn,PuBuGn,PuBu,BuPu,RdPu,PuRd,OrRd,YlOrRd,YlOrBr,Purples,Blues,Greens,Oranges,Reds,Greys,PuOr,BrBG,PRGn,PiYG,RdBu,RdGy,RdYlBu,Spectral,RdYlGn,Accent,Dark2,Paired,Pastel1,Pastel2,Set1,Set2,Set3,Default,Basic,Light,Favorite,Pastel,Balanced,Rpalette,RlatticeBackground,RlatticeShingle,RlatticeLine,RlatticePolygon,Bootstrap,Economist,EconomistBG,Excel,Excel2,Excel3,GGPlot,Solarized,SolarizedBase,PaulTol,ColorBlind,Tableau,TableauGrey,TableauColorBlind,TableauTrafficLight,TableauPurpleGrey,TableauBlueRed,TableauGreenOrange,TableauCyclic,TableauPairSequential,TableauTripleDiverging,WallStreetJournal,WallStreetJournal2,WallStreetJournal3,WallStreetJournalRedGreen,WallStreetJournalBlackGreen,WallStreetJournalDemRep,Stata,Stata2,Stata3,StataMono,BlackAndWhite,CanvasXpress,CanvasXpressT], Default Value: 'User'
-smpTitle: Description: 'Title for the sample axis in one dimensional plots', Type: 'string', Category: 'Samples', Default Value: 'False'
-smpTitleColor: Description: 'Color for the text in sample title', Type: 'color', Category: 'Samples', Default Value: 'rgb(0,0,0)'
-smpTitleScaleFontFactor: Description: 'Scaling factor used to increase or decrease the sample title font size in the canvas', Type: 'float', Category: 'Samples', Default Value: '1'
[…]
