### 2024-11-11 0531ba4
Initial checkin of script to filter the all_results.txt file to only the ones that didn't have perfect similarity score
