### 2024-11-11 84e0ad0
Changed to specific, older version of FlagEmbedding (they released a new version recently that doesn't work with my code anymore)
