### 2024-11-11 b49622e
Modified to allow numeric values to be considered equivalent if they are within 0.5 of each other, at Isaac's request, for json_similarity
