### 2024-11-18 #7328 @luisaforozco
@crvernon thanks for the invitation. This field is outside of my expertise and I do not have access to CanvasXpress, so I will have to pass on reviewing this one.
