### 2024-11-19 #7328 @RahulSundar
> 👋 @RahulSundar - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> Note: this review requires access to CanvasXpress

Hi @crvernon ! Yes, I would like to review this work.
