### 2024-11-25 #7529 @editorialbot
[REVIEW]: Generating Visualizations Conversationally using LLMs
**Submitting author:** @buddyroo30 (<a href="http://orcid.org/0009-0009-6515-1671">Andrew Smith</a>)
**Repository:** https://github.com/buddyroo30/canvasxpress_gen
**Branch with paper.md** (empty if default branch): main
**Version:** v1.1.1
