### 2024-11-25 #7529 @buddyroo30
Hi,
Thank-you for agreeing to go ahead and review our paper and I will work to reduce the word count. However, it has been awhile since we originally submitted the paper and it was uncertain that you were actually going to review it (due to the license issues around CanvasXpress itself). Since submitting we have actually made additional significant progress on this project, and if possible we'd like to update the paper to include our updated content and results. So I'm thinking I'll do that as part of reducing the word count, and I hope that is okay --- but please let me know if you have any concerns.
best regards,Andrew
    On Monday, November 25, 2024 at 11:12:14 AM EST, Chris Vernon ***@***.***> wrote:

👋 @buddyroo30 - as an initial step, please reduce the word count of your paper to around 1000 words max. Thanks!
