### 2024-11-25 #7529 @buddyroo30
Great, thanks. I am out the rest of this week on vacation, but will work on this next week and get you the updated paper soon.
    On Monday, November 25, 2024 at 06:39:09 PM EST, Chris Vernon ***@***.***> wrote:

A-OK @buddyroo30
