### 2024-12-06 7e28c18
Went back to most recent release of FlagEmbedding
