+++ Dockerfile
@@ -4,7 +4,7 @@ ENV DEBIAN_FRONTEND=noninteractive
 RUN apt-get update && apt-get upgrade -y
 RUN apt-get install less emacs nano vim zip unzip curl wget jq gcc libldap2-dev libsasl2-dev -y
 RUN apt-get install sqlite3 -y
-RUN pip install FlagEmbedding==1.2.11
+RUN pip install FlagEmbedding
 RUN mkdir -p /app
 WORKDIR /app
 COPY . /app
