### 2024-12-09 3fd8cdc
Changed lenient equality check from 0.5 to 1.0 as suggested by Isaac, added code to replce __LCB__ and __RCB__ in prompt with { and } --- would consider as interpolation variables otherwise
