### 2024-12-09 5ef1115
A lot of more instructions added by Isaac
