+++ synth_examples/prompt.md
@@ -6,8 +6,38 @@ To help you in this task you are provided detailed information about the valid f
 Adhere to these rules:
 - **Deliberately go through the English text describing a CanvasXpress configuration, the headers or columns names of the data to be visualized, and the detailed valid field information and examples word by word** to appropriately answer the question.
 - **Do not give incorrect responses**. If you cannot generate a CanvasXpress JSON configuration for the given English text and headers/column names, then do not return a response (i.e. simply return empty string '').
+- **Only use valid fields in the JSON you generate!!** The valid CanvasXpress JSON configuration fields are defined below and you can ONLY choose fields from among them; do not make up or hallucinate invalid field names!
 - **Filter out incorrect responses**. Do basic checks on your generated CanvasXpress JSON configuration for errors, for example check if field names are used that are not present in the given valid field information. If you find errors either correct them or do not return a response (i.e. simply return empty string '').
-- **Very very important, return only JSON, do not return anything else!!!** Do not return 3 backticks, or the word json, or any other extraneous text, simply return the generated CanvasXpress configuration JSON.
+- **Very very important, return only JSON, do not return anything else!!!** Do not return 3 backticks, or the word json, or any other extraneous text, simply return the generated CanvasXpress configuration JSON. What you return must be valid JSON.
+- **Only select from valid values of graphType**
+- **Use setMaxX or setMinX if there is only xAxis in the JSON configuration**
+- **Don't use setMaxY if there is no yAxis in the JSON configuration. Use setMaxX instead despite the orientation of the graph.**
+- **Don't use setMinY if there is no yAxis in the JSON configuration. Use setMinX instead despite the orientation of the graph.**
+- **Use "guess" always as the first parameter when including filterData in the JSON configuration**
+- **Use either "like" or "different" always as the third parameter when including filterData in the JSON configuration**
+- **Only select from valid values of graphType**
+- **Never include "orientation" as a parameter when decorations is included in the JSON configuration**
+- **Never include "width" as a parameter when decorations is included in the JSON configuration**
+- **Never include "type" as a parameter when decorations is included in the JSON configuration**
+- **Never include arrays as values for the "x" and "y" and "value" keys in the decorations object**
+- **Always include a single scalar value for the "value" key in the decorations object**
+- **Always include a single scalar value for the "x" key in the decorations object**
+- **Always include a single scalar value for the "y" key in the decorations object**
+- **Only include eitherthe key "x", "y" or "value' in the decorations object, not multiple keys**
+- **If the "x" key is included in the decorations object, do not include the keys "y" and "value" in the JSON configuration**
+- **If the "y" key is included in the decorations object, do not include the keys "x" and "value" in the JSON configuration**
+- **If the "value" key is included in the decorations object, do not include the keys "x" and "y" in the JSON configuration**
+- **If xAxis and yAxis are both included in the JSON configuration, always list xAxis first and then yAxis**
+- **Never include "value" in the configuration when decorations are in the JSON configuration and the graphType is either "Bin", "Binplot", "CDF", "Contour", "Density", "Hex", "Hexplot", "Histogram", "KaplanMeier", "QQ", "Quantile", "Ridgeline", "Scatter2D", "ScatterBubble2D" or "Streamgraph". Always include either "x" or "y"**
[…]
