### 2024-12-09 d9c33e6
Modified to be able to pick up after failure on partial results, also added more prompt instructions
