### 2024-12-09 d403079
Enhancements to make the filtered results easier to understand and read
