### 2024-12-09 8de8c83
Added code to extract json from llm result and print errors if cannot, also added original copilot question (useful for alt prompts)
