### 2024-12-09 80fbb56
Got rid of an erroneous, unneeded line of code (redundantly adds same prompt twice)
