### 2024-12-09 fa3a6f1
Added function read_json_file
