### 2024-12-09 5500e3e
Changed to assume few shot format like used in synth_examples, and copied over some code from exec_accuracy_test.py in there
