+++ vectorize_schema_few_shots.py
@@ -3,8 +3,10 @@ import json
 import sys
 from pymilvus import MilvusClient
 from pymilvus.model.hybrid import BGEM3EmbeddingFunction
+from pymilvus import connections
 import numpy as np
 import os
+import utils
 
 devFlag = os.environ.get("DEV")
 if devFlag == 'True':
@@ -19,6 +21,12 @@ if devFlag:
     allFewShotsFile = "all_few_shots_dev.json"
     cxDocFile = "doc_dev.json"
     vectorDbFile = "/root/.cache/canvasxpress_llm_dev.db"
+
+bge_m3_ef = BGEM3EmbeddingFunction(
+                                    model_name='BAAI/bge-m3', # Specify the model name
+                                    device='cpu', # Specify the device to use, e.g., 'cpu' or 'cuda:0'
+                                    use_fp16=False # Specify whether to use fp16. Set to `False` if `device` is `cpu`.
+                                )
 
 def convert_boolean_dict_values(d):
     for k, v in d.items():
@@ -84,54 +92,60 @@ def generateSchemaRecs(cxConfigInfo):
 
     return(schemaRecs)
 
-bge_m3_ef = BGEM3EmbeddingFunction(
-    model_name='BAAI/bge-m3', # Specify the model name
-    device='cpu', # Specify the device to use, e.g., 'cpu' or 'cuda:0'
-    use_fp16=False # Specify whether to use fp16. Set to `False` if `device` is `cpu`.
-)
+def gen_vectordb(train_docs, vectorDbFile):
+    """
+    Generate a vector database from the training documents.
 
-client = MilvusClient(vectorDbFile)
-client.create_collection(
-    collection_name="few_shot_examples",
-    dimension=1024  # The vectors we will use have 1024 dimensions
-)
+    Args:
+    - train_docs (dict): The training documents.
+    - vectorDbFile (str): The path to the vector database file.
+
+    Returns:
+    - None
+    """
+
+    connections.disconnect("default")
+    if os.path.exists(vectorDbFile):
+        os.remove(vectorDbFile)
+
+    client = MilvusClient(vectorDbFile)
+    client.create_collection(
+        collection_name="few_shot_examples",
+        dimension=1024  # The vectors we will use have 1024 dimensions
+    )
+
+    for i in range(len(train_docs)):
+        curRec = train_docs[i]
+        curConfig = curRec['config']
+        curHeader = curRec['header']
+        curPrompt = curRec['prompt']
+        few_shot_docs_to_embed = []
+        alt_prompts = curRec['alt_prompts']
+        all_prompts = alt_prompts + [curPrompt]
+        for j in range(len(all_prompts)):
+            cur_prompt = all_prompts[j]
+            few_shot_docs_to_embed.append(cur_prompt)
+
+        few_shot_docs_embeddings = bge_m3_ef.encode_documents(few_shot_docs_to_embed)
+
+        for j in range(len(all_prompts)):
+            cur_prompt = all_prompts[j]
+            curDocsBatch = [ {"id": i * 4 + j, "vector": few_shot_docs_embeddings["dense"][j], "configEnglish": cur_prompt, "headers": json.dumps(curHeader), "config": json.dumps(curConfig) } ]
+            few_shot_res = client.insert(
+                collection_name="few_shot_examples",
+                data=curDocsBatch
+            )
+
+    return(client)
+
+vectordb_docs = utils.read_json_file(allFewShotsFile)
[…]
