### 2024-12-25 #7529 @buddyroo30
Hi,
As I mentioned before, we are reworking our JOSS paper to add some improvements and additional testing we have done. I have been busy with other things recently (end of year at work), but we are doing final tests now and I will plan to rework the paper in early January after the holidays, sorry for the delay. I will let you know again when it is ready. Thanks, and happy holidays!
best regards,Andrew Smith
    On Tuesday, December 24, 2024 at 09:44:37 AM EST, Chris Vernon ***@***.***> wrote:

👋 @buddyroo30, @ahmadawais, and @RahulSundar - just checking in to see how things are going with this review. Please post a short update here when you have a moment. Thanks!
