### 2025-01-06 ce35065
Updated rules from Isaac
