### 2025-01-07 e983fe1
Modifed json_similarity to handle more cases, when both are same type int/int or float/float
