### 2025-01-08 d50fb67
Adding directory for updated JOSS paper content
