+++ paper/paper.bib
@@ -1,0 +1,228 @@
+@misc{schulhoff_showing_2024,
+	title = {Showing {Examples}},
+	url = {https://learnprompting.org/docs/basics/few_shot},
+	journal = {Showing Examples},
+	author = {Schulhoff, Sander},
+	month = jul,
+	year = {2024},
+}
+
+@misc{neuhaus_canvasxpress_nodate,
+	title = {{CanvasXpress}: {A} {JavaScript} {Library} for {Data} {Analytics} with {Full} {Audit} {Trail} {Capabilities}},
+	url = {https://www.canvasxpress.org/},
+	author = {Neuhaus, Isaac},
+}
+
+@misc{noauthor_better_nodate,
+	title = {Better language models and their implications},
+	url = {https://openai.com/index/better-language-models/},
+	abstract = {We’ve trained a large-scale unsupervised language model which generates coherent paragraphs of text, achieves state-of-the-art performance on many language modeling benchmarks, and performs rudimentary reading comprehension, machine translation, question answering, and summarization—all without task-specific training.},
+	language = {en-US},
+	urldate = {2024-08-21},
+	file = {Snapshot:C\:\\Users\\smitha26\\Zotero\\storage\\VBA856TC\\better-language-models.html:text/html},
+}
+
+@misc{gao_retrieval-augmented_2023,
+	title = {Retrieval-{Augmented} {Generation} for {Large} {Language} {Models}: {A} {Survey}},
+	shorttitle = {Retrieval-{Augmented} {Generation} for {Large} {Language} {Models}},
+	url = {https://arxiv.org/abs/2312.10997v5},
+	abstract = {Large Language Models (LLMs) showcase impressive capabilities but encounter challenges like hallucination, outdated knowledge, and non-transparent, untraceable reasoning processes. Retrieval-Augmented Generation (RAG) has emerged as a promising solution by incorporating knowledge from external databases. This enhances the accuracy and credibility of the generation, particularly for knowledge-intensive tasks, and allows for continuous knowledge updates and integration of domain-specific information. RAG synergistically merges LLMs' intrinsic knowledge with the vast, dynamic repositories of external databases. This comprehensive review paper offers a detailed examination of the progression of RAG paradigms, encompassing the Naive RAG, the Advanced RAG, and the Modular RAG. It meticulously scrutinizes the tripartite foundation of RAG frameworks, which includes the retrieval, the generation and the augmentation techniques. The paper highlights the state-of-the-art technologies embedded in each of these critical components, providing a profound understanding of the advancements in RAG systems. Furthermore, this paper introduces up-to-date evaluation framework and benchmark. At the end, this article delineates the challenges currently faced and points out prospective avenues for research and development.},
+	language = {en},
+	urldate = {2024-08-21},
+	journal = {arXiv.org},
+	author = {Gao, Yunfan and Xiong, Yun and Gao, Xinyu and Jia, Kangxiang and Pan, Jinliu and Bi, Yuxi and Dai, Yi and Sun, Jiawei and Wang, Meng and Wang, Haofen},
+	month = dec,
+	year = {2023},
+	file = {Full Text PDF:C\:\\Users\\smitha26\\Zotero\\storage\\YVVAXCAB\\Gao et al. - 2023 - Retrieval-Augmented Generation for Large Language .pdf:application/pdf},
+}
+
+@article{li2024longcontextllmsstrugglelong,
+      title={Long-context LLMs Struggle with Long In-context Learning}, 
+      author={Tianle Li and Ge Zhang and Quy Duc Do and Xiang Yue and Wenhu Chen},
+      year={2024},
+      eprint={2404.02060},
+      archivePrefix={arXiv},
+      primaryClass={cs.CL},
[…]
