### 2025-01-21 6834f11
Some additional instructions Isaac gave me to add
