+++ paper/paper.bib
@@ -226,3 +226,14 @@
 	year = {1984},
 	pages = {629--629},
 }
+
+@misc{noauthor_defog-aisqlcoder_nodate,
+	title = {defog-ai/sqlcoder: {SoTA} {LLM} for converting natural language questions to {SQL} queries},
+	shorttitle = {defog-ai/sqlcoder},
+	url = {https://github.com/defog-ai/sqlcoder},
+	abstract = {SoTA LLM for converting natural language questions to SQL queries - defog-ai/sqlcoder},
+	language = {en},
+	urldate = {2025-01-22},
+	journal = {GitHub},
+	file = {Snapshot:C\:\\Users\\smitha26\\Zotero\\storage\\6AMHKNGZ\\sqlcoder.html:text/html},
+}
