### 2025-01-22 df65258
Added sentence about plans to develop agentic component
