+++ paper/paper.md
@@ -66,7 +66,7 @@ We assess accuracy of LLM-generated CanvasXpress JSON configuration using two ke
 
 While accuracy is excellent (close to perfect) for all three datasets, more notably, accuracy is better with datasets containing sentences of up to four words compared to those with six or ten words, with the max length ten dataset showing a more noticeable dropoff in accuracy from six (compared to four to six). This suggests that longer, more complex descriptions of visualizations pose greater challenges for LLMs to generate accurately. How generalizable is this finding (e.g. would it similarly effect other chatbots such as against databases?) And how could it be addressed? We hope to explore this more in the future, but here are some ideas. One potential solution could be to employ a divide-and-conquer approach, constructing a final answer in more precise, incremental LLM generation steps rather than attempting to generate it all at once. “Chain of Thought” prompting seems related to this and might also give improved results.
 
-Looking ahead, we plan to develop a fine-tuned LLM using synthetic data generated by the Copilot. Fine-tuning necessitates at least thousands of examples, which can be effortlessly generated by the copilot, while manually creating such a large dataset is challenging. The potential for improved accuracy through fine-tuning is significant, for example, SQLcoder achieves 93% accuracy compared to GPT-4's 86% in generating SQL from English text [@noauthor_defog-aisqlcoder_nodate].
+Looking ahead, we plan to develop a fine-tuned LLM using synthetic data generated by the Copilot. Fine-tuning necessitates at least thousands of examples, which can be effortlessly generated by the copilot, while manually creating such a large dataset is challenging. The potential for improved accuracy through fine-tuning is significant, for example, SQLcoder achieves 93% accuracy compared to GPT-4's 86% in generating SQL from English text [@noauthor_defog-aisqlcoder_nodate]. We also hope to create an agentic component that can work with other agents and services to support visualization needs.
 
 # User Interface (UI) Integrated into CanvasXpress
 
