### 2025-01-22 #7529 @buddyroo30
Hello, sorry for the delay, but we have now created an updated version of our JOSS submission that we hope you can now begin reviewing, I put it in a new branch JOSS-submission2:
GitHub - buddyroo30/canvasxpress_gen at JOSS-submission2

|
|
|
|  |  |

 |

 |
|
|  |
GitHub - buddyroo30/canvasxpress_gen at JOSS-submission2

Generating CanvasXpress visualizations from natural language descriptions of them using LLMs - GitHub - buddyroo...
 |

 |

 |

We hope this is sufficient for you to review the paper, but please let us know if there are any remaining issues/todos.
thanks,Andrew Smith    On Wednesday, December 25, 2024 at 01:07:30 PM EST, Andrew Smith ***@***.***> wrote:

  Hi,
As I mentioned before, we are reworking our JOSS paper to add some improvements and additional testing we have done. I have been busy with other things recently (end of year at work), but we are doing final tests now and I will plan to rework the paper in early January after the holidays, sorry for the delay. I will let you know again when it is ready. Thanks, and happy holidays!
best regards,Andrew Smith
    On Tuesday, December 24, 2024 at 09:44:37 AM EST, Chris Vernon ***@***.***> wrote:

👋 @buddyroo30, @ahmadawais, and @RahulSundar - just checking in to see how things are going with this review. Please post a short update here when you have a moment. Thanks!
