### 2025-02-27 #7529 @xavieryao
@crvernon  Happy to review here.
