### 2025-03-04 01eb054
Propagating changes to prompt and doc done for accuracy tests for JOSS submission into the main folder for the Flask app to use
