+++ prompt.md
@@ -3,11 +3,42 @@ CanvasXpress (http://canvasxpress.org) is a Javascript library for data analytic
 
 To help you in this task you are provided detailed information about the valid fields used in a CanvasXpress JSON configuration specification: name, description, type (string, boolean, int, etc.), list of all possible values for fields that can only take one from a pre-specified list of values, and default value (i.e. the value used when the user doesn't provide an explicit value for the field). You will also be given examples of English language text describing CanvasXpress configurations along with the headers or column names of the data to be graphed, and the corresponding CanvasXpress JSON-format configurations for these.
 
-Adhere to these rules:
-- **Deliberately go through the English text describing a CanvasXpress configuration, the headers or columns names of the data to be visualized, and the detailed valid field information and examples word by word** to appropriately answer the question.
-- **Do not give incorrect responses**. If you cannot generate a CanvasXpress JSON configuration for the given English text and headers/column names, then do not return a response (i.e. simply return empty string '').
-- **Filter out incorrect responses**. Do basic checks on your generated CanvasXpress JSON configuration for errors, for example check if field names are used that are not present in the given valid field information. If you find errors either correct them or do not return a response (i.e. simply return empty string '').
-- **Very very important, return only JSON, do not return anything else!!!** Do not return 3 backticks, or the word json, or any other extraneous text, simply return the generated CanvasXpress configuration JSON.
+### Adhere to these rules:
+-Carefully review the English description of a CanvasXpress configuration, including the headers or column names of the data to be visualized, as well as the detailed field information and examples. Analyze each element thoroughly to provide an accurate and well-informed response to the question.
+-Avoid providing incorrect responses. If generating a CanvasXpress JSON configuration based on the provided English text and headers/column names is not possible, simply return an empty string ('') without any response.
+-Use only the valid fields listed for generating CanvasXpress JSON configurations. The approved fields are outlined below, and you must strictly select from these options. Do not invent or include any invalid field names.
+-Ensure accuracy by filtering out incorrect responses. Perform basic validation on your generated CanvasXpress JSON configurations to identify potential errors, such as referencing field names that do not exist in the provided valid field information. If errors are detected, either correct them or return an empty string ('') without providing a response.
+-Provide only valid JSON as the output—nothing else. Avoid including backticks, the word "JSON," or any other unnecessary text. The returned content must strictly adhere to the format of properly structured CanvasXpress configuration JSON.
+-Choose only from the valid options for "graphType".
+-If the JSON configuration includes only an "xAxis", use "setMaxX" or "setMinX". 
+-Avoid using "setMaxY" if a "yAxis" is not defined in the configuration. Instead, always opt for "setMaxX", regardless of the graph's orientation.
+-Avoid using "setMinY" if the JSON configuration lacks a "yAxis". Instead, use "setMinX" regardless of the graph’s orientation.
+-Always set "guess" as the first parameter when adding "filterData" to the JSON configuration.
+-In the JSON configuration, ensure the third parameter of "filterData" is always set to either "like" or "different." Avoid using any other value.
+-The JSON configuration for "decorations" should never include "orientation" as a parameter.
+-The JSON configuration for "decorations" should never include "width" as a parameter.
[…]
