### 2025-03-04 5b7c6ef
Updated based on latest few shot examples and doc.json
