+++ schema.txt
@@ -1,42 +1,29 @@
-Here is detailed information about the CanvasXpress JSON configuration fields:
-graphType: Description: 'Specifies the type of graph', Type: 'string', Category: 'General', Options for Field Value: [Bar,Line,Area,AreaLine,BarLine,Boxplot,Dotplot,DotLine,Fish,Gantt,Meter,Heatmap,Stacked,StackedLine,StackedPercent,StackedPercentLine,Tree,Treemap,TagCloud,ParallelCoordinates,Sankey,Scatter2D,ScatterBubble2D,Scatter3D,Correlation,Pie,Venn,Network,Genome,Circular], Default Value: 'Bar'
-graphOrientation: Description: 'Specifies the orientation of one dimensional graphs', Type: 'string', Category: 'General', Options for Field Value: [horizontal,vertical], Default Value: 'horizontal'
-widthFactor: Description: 'Factor to adjust the width of the bars in the one dimensional and the three dimensional plots. The greater the number, the wider the bars', Type: 'integer', Category: 'Aspect Ratio, Space and width', Default Value: '1'
-dataPointSizeScaleFactor: Description: 'Factor used to adjust the size of the dataPointSize', Type: 'float', Category: 'Aspect Ratio, Space and width', Default Value: '1'
-topoJSON: Description: 'TopoJSON map string', Type: 'string', Category: 'Maps', Default Value: 'False'
-mapId: Description: 'Map Id for the map div in the DOM', Type: 'string', Category: 'Maps', Default Value: 'False'
-mapConfig: Description: 'Map config', Type: 'object', Category: 'Maps', Default Value: 'False'
-useLeaflet: Description: 'Use leaflet', Type: 'boolean', Category: 'Maps Leaflet', Default Value: 'False'
-fontName: Description: 'Font name. Web-Safe HTML fonts', Type: 'string', Category: 'Text', Options for Field Value: [American Typewriter,Andale Mono,Arial,Baskerville,Bradley Hand,Brush Script MT,Comic Sans MS,Courier,Didot,Georgia,Gill Sans,Helvetica,Impact,Lucida Console,Luminari,Monaco,Optima,Palatino,Roboto,Tahoma,Times New Roman,Trebuchet MS,Verdana], Default Value: 'Arial'
+graphType: Description: 'Specifies the type of graph to be rendered', Type: 'string', Category: 'General', Options for Field Value: [Area,AreaLine,Bar,BarLine,Boxplot,Bin,Binplot,Bubble,CDF,Chord,Circular,Contour,Correlation,Density,Donut,DotLine,Dotplot,Fish,Gantt,Genome,Heatmap,Hex,Hexplot,Histogram,KaplanMeier,Line,Lollipop,Map,Meter,Network,ParallelCoordinates,Pareto,Pie,QQ,Quantile,Radar,Ridgeline,Sankey,Scatter2D,Scatter3D,ScatterBubble2D,Spaghetti,Stacked,StackedLine,StackedPercent,StackedPercentLine,Streamgraph,Sunburst,TagCloud,Tornado,Tree,Treemap,Upset,Violin,Volcano,Venn,Waterfall], Default Value: 'Bar'
+graphOrientation: Description: 'Specifies the orientation of one-dimensional graphs', Type: 'string', Category: 'General', Options for Field Value: [horizontal,vertical], Default Value: 'horizontal'
[…]
