### 2025-03-04 bb05cc6
Update README.md
Fixed some issues with formatting, adding `` around Makefile targets
