+++ README.md
@@ -9,16 +9,16 @@ The code in this repository can be used to conversationally generate [CanvasXpre
 
 The application is implemented and run using Docker, with various Makefile targets to support building the Docker image, running it, and stopping it. Both a production and development image/application are supported by the Makefile so you can test changes on the development system before propagating them to the production system. The key file artifacts of the system are the high-level prompt file prompt.md (and prompt_dev.md for dev), the schema information or details of the legal/valid CanvasXpress fields in file schema.txt (and schema_dev.txt for dev), and the few shot examples that are stored in the vector database in file all_few_shots.json (all_few_shots_dev.json for dev). app.py contains the Flask application with the ask endpoint taking a user's English description and converting it into a CanvasXpress configuration JSON object. Here are the steps you can take to build and run the image:
 
-1. make build - build the Docker image for the production application.
-2. make build_vector_db - Create the vector database used in the production application for RAG by indexing the few shot examples (the vector database is stored on disk at ~/.cache/canvasxpress_llm.db)
-3. Make build_schema_context - Create the schema.txt file used in the production application, containing the schema info for all the fields present in the few shot examples file (uses file doc.json for the schema info details)
-4. make runi - run the production application interactively (i.e. you will be able to see all the Flask output, including any errors which is useful for debugging, and can exit by typing CTRL-c)
-5. make run - run the production application as a daemon
-6. make exit - shut down the production application
+1. `make build` - build the Docker image for the production application.
+2. `make build_vector_db` - Create the vector database used in the production application for RAG by indexing the few shot examples (the vector database is stored on disk at ~/.cache/canvasxpress_llm.db)
+3. `make build_schema_context` - Create the schema.txt file used in the production application, containing the schema info for all the fields present in the few shot examples file (uses file doc.json for the schema info details)
+4. `make runi` - run the production application interactively (i.e. you will be able to see all the Flask output, including any errors which is useful for debugging, and can exit by typing CTRL-c)
+5. `make run` - run the production application as a daemon
+6. `make exit` - shut down the production application
 
-There is also 'make shell' to enter into a Shell session for the Docker image and 'make buildfresh' which is the same as 'make build' except it doesn't use Docker cache.
+There is also `make shell` to enter into a Shell session for the Docker image and `make buildfresh` which is the same as 'make build' except it doesn't use Docker cache.
 
-To build and run the development version of the application there are Makefile targets corresponding to all the above by simply appending '_dev' (i.e. 'make build_dev', 'make run_dev', 'make exit_dev', etc.) Note that in the provided Makefile the production application runs on port 5008 and the development application runs on port 5009 --- edit RUN_ARGS or RUN_ARGS_DEV to change the ports if you want. Also, the running Docker containers bind mount ~/.cache to /root/.cache inside the container, so make sure you have a .cache directory in your home directory (this is used to store the [BGE-M3](https://milvus.io/docs/embed-with-bgm-m3.md) embedding model files and also the created vector database which is named canvasxpress_llm.db or canvasxpress_llm_dev.db). Also,  ~/.aws/credentials will be bind mounted to /root/.aws/credentials --- this is to support use of models in AWS Bedrock if you want to use them (not needed otherwise).
[…]
