### 2025-03-04 35360a8
Modified from openai to azure_openai (since that is more descriptive of what I am using; I will later add a handler for basic openai)
