### 2025-03-05 d82b9af
Bugfix for build_schema_context_dev --- was using prod image so changed to use dev
