### 2025-03-05 1c4f0ed
Modified so schema.txt and schema_dev.txt get output to the ~/.cache dir
