### 2025-03-07 a00ad72
Modified so details of the available models that can be used is not hardcoded, but stored and read from a separate llm_models.json file
