### 2025-03-07 7c76342
Added handling to support generating results using generic openai (not azure openai)
