### 2025-03-07 0320704
Added import of openai (forgot to add before)
