### 2025-03-23 #7529 @xavieryao
@crvernon @buddyroo30 My apologies on dragging this on for so long. In my view, major revisions are needed.

## Scope & Scholarly Effort
My major concern is on "scholarly effort" and whether this is out-of-scope.

The core of this work is (A) "a broad and illustrative set of few-shot examples" that maps natural language description to CanvasXpress configurations + (B) similarity-based matching to find the few-shot examples similar to user input, and send them as a prompt to LLM.

### Writing prompts / few-shot examples

> While useful for many areas of research, pre-trained machine learning models and notebooks are not in-scope for JOSS.

I'm wondering how does this apply in the age of training-free few-shot / zero-shot / RAG era. Is writing prompt / examples equivalent to pre-training models? @crvernon Your input here is greatly appreciated.

### RAG + few-shot
I appreciate the effort for "thoughtfully curating a broad and illustrative set of few-shot examples". Using RAG to select few-shot examples and construct a prompt is a standard practice ([Example](https://www.googlecloudcommunity.com/gc/Community-Blogs/NLP2SQL-using-dynamic-RAG-based-few-shot-examples/ba-p/805007)).

### CanvasXpress

> submissions that rely upon a proprietary/closed source language or development environment are acceptable. If an open source or free variant of the programming language exists, feel free to encourage the submitting author to consider making their software compatible with the open source/free variant.

This software generates configurations only for CanvasXpress, whose license isn't OSS-compatible. This is acceptable, but I encourage the authors to explore whether the system can be extended to open source/free variant such as Matplotlib / D3 etc.

> that either enables some new research challenges to be addressed or makes addressing research challenges significantly better

Could you add a description of how CanvasXpress is solving research challenges, as an argument that this new system is indirectly solving research challenges?

### Package and Features

> JOSS requires that software should be feature-complete (i.e., no half-baked solutions)

Thumb up / thump down feature has no real effect on the backend. Consider adding some documentation on what's done and what's the plan for future?

> packaged appropriately according to common community standards for the programming language being used

As other reviewers mentioned, right now the files are all in one place. Consider following [this guideline](https://packaging.python.org/en/latest/tutorials/packaging-projects/) to package the software as a Python package.

## Documentation
Please address these items in the [JOSS Checklist](https://joss.readthedocs.io/en/latest/review_checklist.html). There are some tools that could be helpful: [pydoc](https://docs.python.org/3/library/pydoc.html), [unittest](https://docs.python.org/3/library/unittest.html); [GitHub contribution guideline](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Further, consider adding documentation on what features are implemented and what are on-going/placeholders.

## Software paper
* Could you more directly address how this software (through CanvasXpress) solves research problems?
* Could you discuss more on the state of the field? e.g. other RAG+few shot implementation, other plotting software.
* Please directly cite the original source of software / algorithms:

For Milvus:

```bibtex
@inproceedings{2021milvus,
  title={Milvus: A Purpose-Built Vector Data Management System},
  author={Wang, Jianguo and Yi, Xiaomeng and Guo, Rentong and Jin, Hai and Xu, Peng and Li, Shengjun and Wang, Xiangyu and Guo, Xiangzhou and Li, Chengming and Xu, Xiaohai and others},
  booktitle={Proceedings of the 2021 International Conference on Management of Data},
  pages={2614--2627},
  year={2021}
}

@article{2022manu,
  title={Manu: a cloud native vector database management system},
  author={Guo, Rentong and Luan, Xiaofan and Xiang, Long and Yan, Xiao and Yi, Xiaomeng and Luo, Jigao and Cheng, Qianya and Xu, Weizhi and Luo, Jiarui and Liu, Frank and others},
  journal={Proceedings of the VLDB Endowment},
  volume={15},
  number={12},
  pages={3548--3561},
  year={2022},
  publisher={VLDB Endowment}
}
```

For few-shot learning:
https://papers.nips.cc/paper/2020/hash/1457c0d6bfcb4967418bfb8ac142f64a-Abstract.html

* Could you elaborate how the few-shot examples were "thoughtfully created" and quantitatively how broad do they cover?
* Could you give details on "JSON similarity score"?

## The codebase

Will raise issues later to provide suggestions.
