### 2025-04-28 #7529 @buddyroo30
Hi,
No, we have not made any of those changes yet. We sent you our response to the reviews a few weeks and explained there we wanted to get your feedback on that first before proceeding to make the changes to the paper and code (or hear any further feedback or other changes you would want us to make). Please confirm that you received our response to the reviews doc and that you are in agreement with that, then I will go ahead and implement the changes detailed there.
thanks,Andrew
    On Friday, April 25, 2025 at 03:07:45 PM EDT, Rahul Sundar ***@***.***> wrote:

 RahulSundar left a comment (openjournals/joss-reviews#7529)
Dear Authors, @buddyroo30

Have you made the corrections to your codebase and the paper?
Thanks
Rahul
