### 2025-06-10 #7529 @buddyroo30
Hi,
Thanks for the review feedback, so it seems now our plan (detailed in our reply to reviewers a month or so back) is acceptable and we will work now to execute on that. I will keep you updated on our progress. There were a few final notes the reviewers made in these last few emails, and I will write a short reply to those in a follow up reply soon, but just wanted to acknowledge all the reviewers now and let you know we will begin executing our plan.
Thanks,Andrew

Sent from Yahoo Mail for iPad
