### 2025-06-25 2185e66
JOSS revisions: Add experimental results and essential files
Based on synth_data_tests (most recent branch as of 2025-03-07).

New additions:
- Updated .gitignore with OLD_BAK exclusion for cleaner repository
- JOSS_reviews_response.txt with detailed revision plans
- Essential utility scripts and updated configurations
- Complete experimental validation data from synth_examples/

Experimental results include:
- Accuracy tests for 4, 6, and 10 sentence complexity
- All summary results supporting 97%+ accuracy claims
- JOSS submission results and supporting scripts

.gitignore now excludes: OLD_
[…]
