+++ synth_examples/config_to_english/few_shot_examples.txt
@@ -1,0 +1,5 @@
+JSON configuration: { "headers": ["sample","C1","C1","C2"], "configuration": { "graphOrientation":"vertical", "graphType":"Bar", "showLegend":"false", "smpTextRotate":"90", "smpTitle":"Samples", "theme":"CanvasXpress", "title":"Bar Graph Title", "xAxisTitle":"Value" }}, Answer: A bar graph with vertical orientation. The graph type is "Bar" and there is no legend. The sample text is rotated at 90 degrees, and the sample title is "Samples". The theme style used is "CanvasXpress", and the graph has a title "Bar Graph Title". The x-axis title is "Value".
+JSON configuration: { "headers": ["Variable","len","dose","supp","order"], "configuration": { "axisAlgorithm":"rPretty", "axisTextScaleFontFactor":1.8, "axisTitleFontStyle":"bold", "axisTitleScaleFontFactor":1.8, "graphOrientation":"vertical", "graphType":"Boxplot", "groupingFactors":[ "dose" ], "showLegend":"false", "smpTextRotate":"90", "smpTextScaleFontFactor":1.8, "smpTitle":"dose", "smpTitleFontStyle":"bold", "smpTitleScaleFontFactor":1.8, "theme":"CanvasXpress", "title":"The Effect of Vitamin C on Tooth Growth in Guinea Pigs", "xAxis2Show":"false", "xAxisTitle":"len" }}, Answer: A box plot with the following settings: the axis algorithm is set to "rPretty", the axis text font is scaled to 1.8, the axis title font style is bold, the axis title font is scaled to 1.8, the graph orientation is vertical, the graph type is Boxplot, data is grouped by "dose", the legend is not shown, the sample text is rotated 90 degrees, the sample text font is scaled to 1.8, the sample title is "dose", the sample title font style is bold, the sample title font is scaled to 1.8, the theme is "CanvasXpress", the title of the visualization is "The Effect of Vitamin C on Tooth Growth in Guinea Pigs", the secondary x-axis is not shown, and the x-axis title is "len".
+JSON configuration: { "headers": ["Variable","Time1","Time2","Time3","Time4","Time5"], "configuration": { "backgroundType":"panel", "blockContrast":"true", "evenColor":"rgb(226,236,248)", "graphOrientation":"vertical", "graphType":"Line", "legendInside":"true", "legendKeyBackgroundBorderColor":"rgba(255,255,255,0)", "legendKeyBackgroundColor":"rgba(255,255,255,0)", "legendPosition":"topLeft", "panelBackgroundColor":"rgb(226,236,248)", "smpTextRotate":"90", "smpTitle":"Days Old", "theme":"GGPlot", "title":"Growth of Orange Trees", "xAxisTitle":"Circumference (mm)" }}, Answer: A line chart with the following properties: the background type is set to "panel", block contrast is enabled, even color is set to RGB(226, 236, 248), graph orientation is vertical, the graph type is a line chart, the legend is placed inside the chart, legend key background border color is set to RGBA(255, 255, 255, 0), legend key background color is set to RGBA(255, 255, 255, 0), the legend position is set to the top left corner, the panel background color is set to RGB(226, 236, 248), sample text is rotated 90 degrees, sample title is "Days Old", the theme is set to "GGPlot", the chart title is "Growth of Orange Trees", and the x-axis title is "Circumference (mm)".
[…]

[…]
