### 2025-06-26 b1d49a0
Commiting all the code changs that Roocode made in modifying the repo to meet the JOSS reviewer requirents. haven't tested yet, will do that next, then possibly make further changes based on test results.
