+++ paper/paper.md
@@ -1,7 +1,7 @@
 ---
 title: 'Generating Visualizations Conversationally using Guided Autocomplete and LLMs'
 tags:
-  - Artifical Intelligence (AI)
+  - Artificial Intelligence (AI)
   - Natural Language Processing (NLP)
   - Large Language Model (LLM)
   - Generative AI
@@ -36,9 +36,11 @@ CanvasXpress [@neuhaus_canvasxpress_nodate] is a powerful JavaScript library tha
 
 A crucial component of creating visualizations with CanvasXpress is the JSON-format configuration object, which specifies the parameters and features of the visual representation to be generated. While a fundamental configuration can be straightforward, the potential options for customization are numerous, allowing for tailored visual experiences. A basic config might include attributes like chart type (e.g., bar, line, scatter), data source, axis labels, and colors. The data for a CanvasXpress visualization can be given as an "array of arrays", where the first array serves as the column headers and subsequent arrays are the data to be graphed.
 
-While tools like CanvasXpress provide a robust framework for creating dynamic visualizations, the requirement for coding and web development skills can be a barrier for many potential users. It begs the question: can we simplify this process for users by allowing them more easily to create visualizations by simply describing them in plain English? By integrating tools like CanvasXpress into an AI-driven platform and using guided autocomplete, users can potentially generate complex visualizations without any coding knowledge.
+While tools like CanvasXpress provide a robust framework for creating dynamic visualizations, the requirement for coding and web development skills can be a barrier for many potential users. This creates a significant research challenge: many domain scientists in fields like genomics, proteomics, and clinical studies have deep knowledge of their field but limited visualization programming skills, creating a technical barrier to effective data exploration and communication. It begs the question: can we simplify this process for users by allowing them more easily to create visualizations by simply describing them in plain English? By integrating tools like CanvasXpress into an AI-driven platform and using guided autocomplete, users can potentially generate complex visualizations without any coding knowledge, directly addressing this research bottleneck.
 
-We have created such a system leveraging guided autocomplete and LLMs to empower users to easily create powerful visualizations conversationally. To simplify the process of generating visualizations, users can upload a tab or comma-separated file containing both headers and data. This allows for easy integration of data into the CanvasXpress platform. Once the data is uploaded, users can be guided to describe the desired graph in plain English, specifying exactly how they envision the representation of their data. For instance, after uploading a file containing sales data, a user could request, “I want a line graph comparing the quarterly sales growth of Product A and Product B.” [@noauthor_canvasxpress_nodate]
+We have created such a system leveraging guided autocomplete and LLMs to empower users to easily create powerful visualizations conversationally. CanvasXpress has several key attributes that make it particularly suitable for LLM-based visualization generation: (1) Its well-structured JSON schema with semantically meaningful field names creates a natural mapping between English descriptions and configuration elements; (2) Its declarative approach (unlike ggplot2's sequential function calls or D3.js's DOM manipulation) is more compatible with one-shot generation by LLMs; (3) CanvasXpress implements sensible defaults and handles missing parameters gracefully, reducing errors in LLM-generated configurations; (4) The scientific domain focus and balanced configuration complexity make it well-suited for the kinds of visualizations researchers commonly need. While tools like Plotly also use JSON configurations and could potentially be adapted to our approach, many visualization libraries would require generating procedural code rather than declarative configurations, making them significantly more challenging for LLM-based generation.
+
[…]
