+++ src/canvasxpress_gen/rag/__init__.py
@@ -1,0 +1,19 @@
+"""
+RAG (Retrieval Augmented Generation) module for CanvasXpress Generation System.
+
+This module provides vector database operations, embedding services, and retrieval
+functionality for enhancing LLM responses with relevant context from CanvasXpress
+documentation and few-shot examples.
+"""
+
+from .embedding_service import EmbeddingService
+from .vector_store import VectorStore
+from .retrieval_service import RetrievalService
+from .schema_processor import SchemaProcessor
+
+__all__ = [
+    'EmbeddingService',
+    'VectorStore', 
+    'RetrievalService',
+    'SchemaProcessor'
+]
