+++ app_refactored.py
@@ -1,0 +1,515 @@
+"""
+Refactored Flask application for CanvasXpress Generation System.
+
+This version uses the new modular architecture with proper imports and
+professional code organization as required by JOSS reviewers.
+"""
+
+from flask import Flask, render_template, request, redirect, url_for, flash, abort, session, jsonify, send_from_directory, send_file
+from werkzeug.utils import secure_filename
+import os
+import sys
+import re
+import datetime
+import time
+import json
+import urllib.parse
+import boto3
+from tabulate import tabulate
+import numpy as np
+from dotenv import load_dotenv
+from typing import Dict, Any, List, Optional, Tuple
+
+# Import our new modular components
+from canvasxpress_gen.llm import LLMService
+from canvasxpress_gen.rag import RetrievalService
+from canvasxpress_gen.utils import (
+    JSONSimilarity, ConfigValidator, 
+    read_json_file, empty, random_password,
+    clean_llm_response_text, parse_file,
+    getSiteMinderUser, getSMRedirectUrl
+)
+from AESCipher import AESCipher
+
+# Load environment variables
+load_dotenv()
+
+
+class CanvasXpressApp:
+    """
+    Main application class for CanvasXpress Generation System.
+    
+    This class encapsulates the Flask application and provides a clean
+    interface for the CanvasXpress generation functionality.
+    """
+    
+    def __init__(self):
+        """Initialize the CanvasXpress application."""
+        self.app = Flask(__name__)
+        self.dev_flag = self._get_dev_flag()
+        self.sm_validation = self._get_sm_validation()
+        self.validated_cookies = {}
+        
+        # Initialize services
+        self.llm_service = None
+        self.retrieval_service = None
+        self.config_validator = ConfigValidator()
+        
+        # Configuration
+        self._setup_app_config()
+        self._setup_file_paths()
+        self._initialize_services()
+        self._register_routes()
+    
+    def _get_dev_flag(self) -> bool:
+        """Get development flag from environment."""
+        dev_flag = os.environ.get("DEV")
+        return dev_flag == 'True' if dev_flag else False
+    
+    def _get_sm_validation(self) -> bool:
+        """Get SiteMinder validation flag from environment."""
+        sm_val = os.environ.get('SMVAL')
+        return sm_val == 'True' if sm_val else False
+    
+    def _setup_app_config(self) -> None:
+        """Setup Flask application configuration."""
+        self.app.config['SECRET_KEY'] = random_password(16)
+        self.app.config['UPLOAD_FOLDER'] = "/tmp"
+        self.aes = AESCipher(self.app.config['SECRET_KEY'])
+    
+    def _setup_file_paths(self) -> None:
+        """Setup file paths based on development flag."""
+        if self.dev_flag:
+            self.prompt_file = "prompt_dev.md"
+            self.schema_info_file = "/root/.cache/schema_dev.txt"
+            self.docs_file = "doc_dev.json"
+            self.examples_file = "all_few_shots_dev.json"
+        else:
+            self.prompt_file = "prompt.md"
+            self.schema_info_file = "/root/.cache/schema.txt"
+            self.docs_file = "doc.json"
+            self.examples_file = "all_few_shots.json"
+        
+        # Validate required files exist
+        if not os.path.exists(self.schema_info_file):
+            print(f"Schema file not found: {self.schema_info_file}")
+            print("Please generate the schema file first; exiting...")
+            sys.exit(1)
+    
+    def _initialize_services(self) -> None:
+        """Initialize LLM and RAG services."""
+        try:
+            # Initialize LLM service
+            self.llm_service = LLMService()
+            
+            # Load and validate LLM models
+            llm_models_file = "/root/.cache/llm_models.json"
+            if os.path.exists(llm_models_file):
[…]

[…]
