### 2025-06-26 ca45496
Fixes from Roocode
