### 2025-06-26 668ea89
Results of testing and debugging to fix using Roocode, system now seems to work correctly but still some styling/ui issues with the simple demo ui that need to be addressed.
