+++ app_refactored.py
@@ -21,11 +21,11 @@ from dotenv import load_dotenv
 from typing import Dict, Any, List, Optional, Tuple
 
 # Import our new modular components
-from canvasxpress_gen.llm import LLMService
-from canvasxpress_gen.rag import RetrievalService
-from canvasxpress_gen.utils import (
-    JSONSimilarity, ConfigValidator, 
-    read_json_file, empty, random_password,
+from src.canvasxpress_gen.llm import LLMService
+from src.canvasxpress_gen.rag import RetrievalService
+from src.canvasxpress_gen.utils import (
+    JSONSimilarity, ConfigValidator,
+    load_json_file, empty, random_password,
     clean_llm_response_text, parse_file,
     getSiteMinderUser, getSMRedirectUrl
 )
@@ -105,22 +105,33 @@ class CanvasXpressApp:
             # Load and validate LLM models
             llm_models_file = "/root/.cache/llm_models.json"
             if os.path.exists(llm_models_file):
-                self.llm_models = read_json_file(llm_models_file)
+                self.llm_models = load_json_file(llm_models_file)
                 self.llm_models_client = self._convert_model_data(self.llm_models)
             else:
                 print(f"LLM models file not found: {llm_models_file}")
                 print("Please generate the LLM models file first; exiting...")
                 sys.exit(1)
             
-            # Initialize RAG service
-            self.retrieval_service = RetrievalService()
-            if os.path.exists(self.docs_file) or os.path.exists(self.examples_file):
-                self.retrieval_service.initialize(
-                    docs_file=self.docs_file,
-                    examples_file=self.examples_file
+            # Initialize direct PyMilvus connection (same as vectorize_schema_few_shots.py)
+            from pymilvus import MilvusClient
+            from pymilvus.model.hybrid import BGEM3EmbeddingFunction
+            
+            self.vector_db_file = "/root/.cache/canvasxpress_llm_dev.db" if self.dev_flag else "/root/.cache/canvasxpress_llm.db"
+            
+            if os.path.exists(self.vector_db_file):
+                # Direct connection using same approach as vectorize_schema_few_shots.py
+                self.milvus_client = MilvusClient(self.vector_db_file)
+                self.bge_m3_ef = BGEM3EmbeddingFunction(
+                    model_name='BAAI/bge-m3',
+                    device='cpu',
+                    use_fp16=False
                 )
+                print(f"Direct PyMilvus connection established: {self.vector_db_file}")
+                self.retrieval_service = "direct_milvus"  # Flag for direct connection
             else:
-                print("Warning: RAG service could not be fully initialized - missing data files")
+                print(f"Warning: Vector database not found: {self.vector_db_file}")
+                print("Please run 'make build_vector_db' first")
+                self.retrieval_service = None
             
             print("Services initialized successfully")
             
@@ -134,9 +145,9 @@ class CanvasXpressApp:
         client_models = []
         for model_name, model_info in models.items():
             client_model = {
-                'name': model_name,
+                'value': model_name,  # JavaScript expects 'value' property
+                'text': model_info.get('description', model_name),  # JavaScript expects 'text' property
                 'type': model_info.get('type', 'unknown'),
-                'description': model_info.get('description', ''),
                 'provider': model_info.get('provider', 'unknown')
             }
             client_models.append(client_model)
@@ -202,44 +213,111 @@ class CanvasXpressApp:
             format_type = 'text'
         
         try:
-            if self.retrieval_service:
+            if self.retrieval_service == "direct_milvus" and hasattr(self, 'milvus_client'):
                 if num == 'all':
-                    # Get all examples
-                    examples = self.retrieval_service.schema_processor.few_shot_examples
-                    if format_type == 'json':
-                        return json.dumps([{
[…]
