### 2025-06-27 864390a
Roocode added documentation about the simple demo web UI
