### 2025-06-27 c49149e
Added text explaining that guided autocomplete/copilot functionality is actually in the main CanvasXpress repo
