### 2025-06-30 09ec8c2
Various updates by Roocode while trying out the testing framework and fixing issues. It also added, in addition to unit tests, an end-to-end integration test. The tests will use actual vector db and LLM API calls if configured, otherwise backup to using mocks
