+++ pytest.ini
@@ -3,13 +3,16 @@ testpaths = tests
 python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
-addopts = 
+# Exclude problematic directories that contain interactive code
+norecursedirs = OLD_BAK .git __pycache__ .pytest_cache
+addopts =
     --verbose
     --tb=short
     --cov=src/canvasxpress_gen
     --cov-report=html
     --cov-report=term-missing
     --cov-fail-under=80
+    --ignore=OLD_BAK
 markers =
     unit: Unit tests
     integration: Integration tests
