### 2025-06-30 999e911
Modified to not hardcode number of few shot examples, but to read from .env file
