### 2025-06-30 9e38bfb
Updates to docs based on updated testing and ability to set num few shots
