### 2025-06-30 6ffa76e
Some updates to the docs done by Roocode
