### 2025-06-30 64a9b68
Added new text to the paper about Python structure and testing
