+++ paper/paper.md
@@ -76,6 +76,8 @@ Looking ahead, we plan to develop a fine-tuned LLM using synthetic data generate
 
 We have also developed an agentic version of our system using AWS Bedrock, which extends the capabilities beyond simple visualization generation. This agent-based implementation can interact with multiple data sources, perform complex multi-step analysis workflows, and integrate with other AI agents to provide comprehensive data analysis and visualization pipelines. The agent architecture enables more sophisticated use cases, such as automatically suggesting appropriate visualization types based on data characteristics, performing data preprocessing and validation, and generating multiple complementary visualizations for comprehensive data exploration. This agentic approach represents a significant step toward fully autonomous data analysis and visualization systems.
 
+The system is implemented as a professional Python package with modular architecture, featuring comprehensive automated testing with over 85 unit and integration tests that validate both real API functionality and graceful fallback to mock responses when external services are unavailable. This ensures the system works reliably across different deployment environments, from development to production, and provides confidence in system reliability for research applications.
+
 # Research Impact and Applications
 
 Our natural language interface system directly addresses significant research challenges by democratizing access to sophisticated visualization capabilities. By providing a natural language interface to visualization creation, our system: (1) Reduces the time from data analysis to visualization, enabling faster hypothesis generation and testing; (2) Enables researchers to explore more visualization options without technical obstacles, facilitating better data exploration; (3) Eliminates the need for specialized visualization support or collaboration with programmers, making researchers more self-sufficient; (4) Accelerates the research process in data-intensive scientific domains where visualization is critical to discovery and communication of findings. For example, a genomics researcher can quickly generate complex heatmaps to visualize gene expression patterns by simply describing "show me a heatmap of gene expression levels across different tissue types, clustered by similarity" rather than learning CanvasXpress's JSON configuration syntax or requiring programming assistance.
