### 2025-06-30 b9c626b
Had Roocode generate alternative versions of the paper at different word lengths; paper_medium.md goes right up to allowable word count so will probably go with it after review
