### 2025-07-01 fbae34b
Removed medium and short sized papers, they aren't needed (not going with them)
