+++ paper/paper_medium.md
@@ -1,103 +1,0 @@
----
-title: 'Generating Visualizations Conversationally using Guided Autocomplete and LLMs'
-tags:
-  - Artificial Intelligence (AI)
-  - Natural Language Processing (NLP)
-  - Large Language Model (LLM)
-  - Generative AI
-  - Visualization
-  - Data Analytics
-  - Exploratory Data Analysis (EDA)
-  - Bioinformatics
-  - CanvasXpress
-authors:
-  - name: Andrew K Smith
-    orcid: 0009-0009-6515-1671
-    equal-contrib: true
-    affiliation: "1"
-  - name: Isaac Neuhaus
-    orcid: 0000-0002-5622-8683
-    equal-contrib: true
-    affiliation: 1
-affiliations:
- - name: Informatics & Predictive Sciences, Knowledge Science Research, Computational Genomics, Bristol Myers Squibb 3551 Lawrenceville Rd, Lawrence Township, NJ 08648
-   index: 1
-date: 30 September 2024
-bibliography: paper.bib
----
-
-# Summary
-
-Powerful data visualization packages like CanvasXpress provide rich features for exploring datasets but have high learning curves requiring detailed web development skills. We demonstrate how guided autocomplete and Large Language Models (LLMs) can be effectively used together in a complementary way to greatly simplify visualization creation. Users upload datasets as tab-delimited or CSV files and describe desired visualizations in plain English, guided by autocomplete suggestions or through free-form text input. Following guided suggestions generates configurations with perfect accuracy, while LLMs serve as backup with high accuracy when users deviate from suggestions.
-
-The guided autocomplete system can generate synthetic training data at large scale for few-shot prompting or LLM fine-tuning. We evaluate accuracy using thousands of synthetic examples with Retrieval Augmented Generation (RAG), achieving over 97% exact match accuracy across datasets. Results demonstrate that accuracy increases with more few-shot examples but decreases with prompt complexity—testing datasets with maximum 4, 6, and 10 sentences shows that longer, more complex visualization descriptions pose greater challenges for LLMs to handle accurately.
-
-# Statement of Need
-
-CanvasXpress [@neuhaus_canvasxpress_nodate] is a powerful JavaScript library for creating interactive data visualizations using HTML5 canvas technology. It supports various chart types (line, bar, scatter, 3D graphs) with extensive customization options through JSON configuration objects. A crucial component of creating visualizations is the JSON-format configuration object, which specifies parameters and features of the visual representation. While basic configurations can be straightforward, the numerous customization options allow for tailored visual experiences but require technical expertise.
-
-The requirement for coding and web development skills creates significant barriers for many potential users. This presents a research challenge: many domain scientists in fields like genomics, proteomics, and clinical studies have deep knowledge of their field but limited visualization programming skills, creating technical barriers to effective data exploration and communication. Can we simplify this process by enabling users to create visualizations through natural language descriptions?
-
-We created a system leveraging guided autocomplete and LLMs for conversational visualization generation. CanvasXpress has several key attributes making it particularly suitable for LLM-based generation: (1) Its well-structured JSON schema with semantically meaningful field names creates natural mapping between English descriptions and configuration elements; (2) Its declarative approach (unlike ggplot2's sequential function calls or D3.js's DOM manipulation) is more compatible with one-shot generation by LLMs; (3) CanvasXpress implements sensible defaults and handles missing parameters gracefully, reducing errors in LLM-generated configurations; (4) The scientific domain focus and balanced configuration complexity make it well-suited for visualizations researchers commonly need.
-
-While tools like Plotly also use JSON configurations and could potentially be adapted to our approach, many visualization libraries would require generating procedural code rather than declarative configurations, making them significantly more challenging for LLM-based generation.
-
-# Implementation
-
-## Guided Autocomplete
-
[…]
