+++ paper/paper_short.md
@@ -1,83 +1,0 @@
----
-title: 'Generating Visualizations Conversationally using Guided Autocomplete and LLMs'
-tags:
-  - Artificial Intelligence (AI)
-  - Natural Language Processing (NLP)
-  - Large Language Model (LLM)
-  - Generative AI
-  - Visualization
-  - Data Analytics
-  - Exploratory Data Analysis (EDA)
-  - Bioinformatics
-  - CanvasXpress
-authors:
-  - name: Andrew K Smith
-    orcid: 0009-0009-6515-1671
-    equal-contrib: true
-    affiliation: "1"
-  - name: Isaac Neuhaus
-    orcid: 0000-0002-5622-8683
-    equal-contrib: true
-    affiliation: 1
-affiliations:
- - name: Informatics & Predictive Sciences, Knowledge Science Research, Computational Genomics, Bristol Myers Squibb 3551 Lawrenceville Rd, Lawrence Township, NJ 08648
-   index: 1
-date: 30 September 2024
-bibliography: paper.bib
----
-
-# Summary
-
-Data visualization packages like CanvasXpress provide rich features for exploring datasets but have high learning curves requiring web development skills. We demonstrate how guided autocomplete and Large Language Models (LLMs) can be used together to simplify visualization creation. Users upload datasets and describe desired visualizations in plain English, guided by autocomplete suggestions or free-form text. Following guided suggestions generates configurations with perfect accuracy, while LLMs serve as backup with high accuracy. The guided autocomplete system generates synthetic training data at scale for few-shot prompting or LLM fine-tuning. We evaluate accuracy using thousands of synthetic examples with Retrieval Augmented Generation (RAG), achieving over 97% exact match accuracy. Results show accuracy increases with more few-shot examples but decreases with prompt complexity (4 vs 6 vs 10 sentences), demonstrating that complex visualization requests are more challenging for LLMs.
-
-# Statement of Need
-
-CanvasXpress [@neuhaus_canvasxpress_nodate] is a JavaScript library for creating interactive data visualizations using HTML5 canvas technology. It supports various chart types with extensive customization through JSON configuration objects. However, the requirement for coding skills creates barriers for domain scientists in genomics, proteomics, and clinical studies who have deep field knowledge but limited programming skills. This creates a research challenge: can we enable visualization creation through natural language descriptions?
-
-We created a system leveraging guided autocomplete and LLMs for conversational visualization generation. CanvasXpress is particularly suitable for LLM-based generation due to: (1) well-structured JSON schema with semantically meaningful field names; (2) declarative approach compatible with one-shot LLM generation; (3) sensible defaults and graceful handling of missing parameters; (4) scientific domain focus matching researcher needs. Unlike procedural libraries requiring complex code generation, CanvasXpress's declarative configurations provide natural targets for LLM generation.
-
-# Implementation
-
-## Guided Autocomplete
-
-Guided autocomplete helps users build prompts incrementally by suggesting valid completions. Users typing "Create" see options like "a bar graph" or "a line chart." This ensures perfect accuracy when followed completely, with LLMs serving as backup for free-form input. The system, called "CoPilot," draws from menu-based natural language approaches [@tennant_menu-based_1984] and generates synthetic prompt/config pairs for LLM training.
-
-## Prompt Engineering and RAG
-
-We use few-shot prompting [@brown_language_2020; @schulhoff_showing_2024] with examples automatically generated by the guided autocomplete system—not manually curated. This automated approach generates thousands of diverse, high-quality examples efficiently. We provide schema information for ~150 most-used CanvasXpress fields (of ~1500 total) due to context window limits.
-
[…]
