### 2025-07-01 056c23c
Shortened main paper (one for actual submission to JOSS), and created extended version of the paper with more detail
