+++ paper/paper.md
@@ -28,76 +28,60 @@ bibliography: paper.bib
 
 # Summary
 
-Powerful data visualization and analysis packages exist such as CanvasXpress providing rich features for exploring datasets. However, such packages can have a high learning curve and be difficult to use, even requiring detailed web development skills. Here we demonstrate how two techniques, guided autocomplete and Large Language Models (“LLMs”), can be effectively used together in a complementary way to greatly simplify the use of such packages, focusing specifically on CanvasXpress. In our developed system users first upload their dataset as a tab-delimited / CSV file or start with any CanvasXpress visualization and then they are guided to describe in plain English their desired visualization by the system prompting the user to choose from among the legal English continuations at every step. It is not mandatory that the user choose from the guided autocomplete suggestions, they can simply break out and free-form describe their visualization as they please. If the user does follow all the suggestions then the configuration to create their visualization can be created with perfect accuracy, and if not then the LLM serves as a backup and generates with high accuracy the visualization configuration. The guided autocomplete system can also be used to generate at large scale synthetic training data for the LLM, which can be used as a source of "few shot" examples in prompting general LLMs like GPT-4o or as a large enough dataset to fine-tune an LLM to the task of generating CanvasXpress visualization configuration. To demonstrate how the guided autocomplete complements the LLM, and to assess accuracy of LLM generated visualization configuration, we use the guided autocomplete system to generate several thousand synthetic examples, which are used as potential few shot examples for retrieval augmented generation (“RAG”), and assess the accuracy of LLM generated visualization configuration as the number of potential few shot examples increases. We also assess for 3 synthetic datasets of increasing maximum number of sentences per example (4, 6, and 10) to see how accuracy scales as the complexity of the problem increases (i.e. more sentences means more complex). Excellent accuracy (over 97% exactly matching known correct answers) is achieved for all the datasets and increases as the number of potential few shot examples increases, however we do find that accuracy is lower for larger numbers of sentences demonstrating that more complex tasks are more difficult for the LLM to handle correctly.
+Powerful data visualization packages like CanvasXpress offer rich features for exploring datasets but often have a high learning curve, requiring detailed web development skills. We present a system that combines guided autocomplete and Large Language Models (LLMs) to simplify the use of such packages, focusing specifically on CanvasXpress. Users upload tabular data or start from any CanvasXpress visualization, then describe their desired visualization in plain English—either by following guided suggestions or typing free-form text. Guided autocomplete ensures perfect configuration accuracy, while LLMs serve as a robust backup for flexible. free-form input. The system can also generate large-scale synthetic training data for use as few-shot examples or for fine-tuning. We evaluate accuracy of LLM CanvasXpress configuration generation using thousands of synthetic examples with Retrieval Augmented Generation (RAG), achieving over 97% exact match accuracy. Results show that accuracy increases with more few-shot examples but decreases with prompt complexity, highlighting the challenge of generating complex visualizations from natural language.
 
 # Statement of Need
 
-CanvasXpress [@neuhaus_canvasxpress_nodate] is a powerful JavaScript library that allows one to create interactive and visually appealing graphical representations of data. Unlike traditional charting libraries, CanvasXpress employs HTML5 canvas technology, offering excellent performance and flexibility in rendering various types of visualizations, including line charts, bar charts, scatter plots, and 3D  graphs. CanvasXpress supports a range of  customization options, allowing users to adjust colors, fonts, and styles to align with their unique branding requirements. CanvasXpress is an invaluable tool for any data-focused application, bridging the gap between complex data sets and intuitive visual storytelling.
[…]
