### 2025-07-01 8c8d61d
Simplifications of documentation files (markdowns) based on Roocode's suggestions
