+++ docs/API.md
@@ -1,775 +1,238 @@
 # CanvasXpress Generation System API Documentation
 
-This document provides comprehensive documentation for the CanvasXpress Generation System REST API, including endpoint specifications, integration examples, and connection guides.
-
-## Table of Contents
-
-- [Overview](#overview)
-- [Authentication](#authentication)
-- [Base URL and Versioning](#base-url-and-versioning)
-- [API Endpoints](#api-endpoints)
-- [Integration Examples](#integration-examples)
-- [CanvasXpress Integration](#canvasxpress-integration)
-- [Error Handling](#error-handling)
-- [Rate Limiting](#rate-limiting)
-- [SDKs and Libraries](#sdks-and-libraries)
+REST API documentation for the CanvasXpress Generation System backend service.
 
 ## Overview
 
-The CanvasXpress Generation System provides a RESTful API that enables applications to generate CanvasXpress visualizations from natural language descriptions. The API is designed to be simple, reliable, and scalable for both individual and enterprise use.
+The CanvasXpress Generation System provides a RESTful API that enables CanvasXpress to generate visualizations from natural language descriptions. This is a **backend service** designed for integration with CanvasXpress.
 
-### Key Features
+## Base URLs
 
-- **Natural Language Processing**: Convert English descriptions to CanvasXpress configurations
-- **Multi-format Support**: Accept CSV, TSV, and JSON data formats
-- **RAG-Enhanced Generation**: Leverage retrieval-augmented generation for improved accuracy
-- **Multiple LLM Support**: Choose from various language models
-- **Enterprise Security**: SiteMinder SSO and secure authentication options
-
-### API Architecture
-
-```
-┌─────────────────┐    ┌──────────────────┐    ┌─────────────────┐
-│   Client App    │───▶│  Flask API       │───▶│  LLM Service    │
-│                 │    │    (app.py)      │    │                 │
-└─────────────────┘    └──────────────────┘    └─────────────────┘
-                              │                          │
-                              ▼                          ▼
-                       ┌──────────────────┐    ┌─────────────────┐
-                       │  RAG System      │    │  Vector DB      │
-                       │  (Retrieval)     │    │  (Milvus)       │
-                       └──────────────────┘    └─────────────────┘
-```
+- **Production**: `http://localhost:5008/`
+- **Development**: `http://localhost:5009/`
 
 ## Authentication
 
-### SiteMinder SSO (Enterprise)
-
-When SiteMinder is enabled (`SMVAL=True`), all requests require valid SiteMinder session cookies:
-
-```http
-Cookie: SMSESSION=your_session_token
-```
-
-### Development Mode
-
-For development and testing (`SMVAL=False`), no authentication is required.
-
-### API Key Authentication (Future)
-
-API key authentication is planned for future releases:
-
-```http
-Authorization: Bearer your_api_key
-```
-
-## Base URL and Versioning
-
-### Production
-```
-https://your-domain.com/
-```
-
-### Development
-```
-http://localhost:5008/  # Production container
-http://localhost:5009/  # Development container
-```
-
-### API Versioning
-
-Currently, the API is unversioned. Future versions will use URL-based versioning:
-```
-/api/v1/endpoint
-```
+- **Development**: No authentication required
+- **Enterprise**: SiteMinder SSO (`SMVAL=True` requires session cookies)
 
 ## API Endpoints
 
-### 1. Home Page
[…]
