### 2025-07-01 905e2ac
Updates to markdown doc files by Roocode, e.g. to remove references to JOSS reviewers, etc.
