### 2025-07-01 405baa1
Small update to description of simple included web ui
