### 2025-07-02 6b48665
Added simple statement about the web ui, its purpose
