### 2025-07-02 01a7d96
Updated/fixed direct integration example
