### 2025-07-02 c4a9c8e
Code fixes in doc files by Roocode
