### 2025-07-02 0b22528
Modified content about SiteMinder support
