### 2025-07-02 d78df3a
Removed CORS section
