### 2025-07-02 c2ccfc2
Added notes about being able to access over internet, not just localhost
