### 2025-07-02 3435d6c
Reduced emphasis on SiteMinder
