### 2025-07-02 3cb7423
Fixes by Roocode to main API endpoint ask
