### 2025-07-02 f91ddf6
Fixes to give more details of response format
