### 2025-07-02 b6a6704
Added more detail for get_few_shots, and removed detailed integration section which is already in INTEGRATION.md
