### 2025-07-02 14330df
Updated available models section
