+++ docs/API.md
@@ -221,14 +221,19 @@ SMVAL=False
 
 ### Available Models
 Configure in `llm_models.json`:
+
 ```json
 {
-  "gemini-1.5-flash": {"type": "google_gemini", "provider": "google"},
-  "gpt-4o": {"type": "openai", "provider": "openai"},
-  "gpt-4o-global": {"type": "azure_openai", "provider": "azure"}
+  "gemini-1.5-flash": {"type": "google_gemini", "text": "Gemini 1.5 Flash"},
+  "gpt-4o-global": {"type": "azure_openai", "text": "GPT-4o"},
+  "anthropic.claude-3-5-sonnet-20240620-v1:0": {"type": "anthropic", "text": "Anthropic Claude Sonnet 3.5"},
+  "mistral.mistral-large-2407-v1:0": {"type": "mistral", "text": "Mistral Large 2 (24.07)"},
+  "meta.llama3-1-70b-instruct-v1:0": {"type": "llama31", "text": "Llama 3.1 70B"},
+  "gpt-4-32k": {"type": "azure_openai", "text": "GPT-4-32k"}
 }
 ```
 
+*Note: This shows a subset of available models. See `llm_models.json` for the complete list.*
 ## Development vs Production Interface
 
 ### Development Interface
