### 2025-07-02 e5a3abf
more fixes to available models
