### 2025-07-02 b715524
More fixes to available models
