### 2025-07-02 244c244
updated to say tests must be done in docker container
