+++ CONTRIBUTING.md
@@ -1,30 +1,18 @@
 # Contributing to CanvasXpress Generation System
 
-Thank you for your interest in contributing! This guide provides information about how others can contribute to the project, report issues, and seek support.
+Thank you for considering contributing to the CanvasXpress Generation System! 🎉
 
-## 🚀 Quick Setup
+This guide provides information about how others can contribute to the project, report issues, and seek support.
 
-### Prerequisites
-- Python 3.9+, Docker, Git
-- LLM API access (OpenAI, Google, AWS Bedrock, or Ollama)
+## 🚀 Getting Started
 
-### Development Environment
-```bash
-git clone https://github.com/buddyroo30/canvasxpress_gen.git
-cd canvasxpress_gen
-python -m venv venv && source venv/bin/activate
-pip install -r requirements.txt
+**First-time setup:** Follow the complete setup instructions in [README.md](README.md)
 
-# Set up environment
-cp .env.example .env  # Add your API keys
-make build_dev
-make build_vector_db_dev
-make run_dev  # Runs on port 5009
-```
+**For development:** Use the development setup and commands detailed in the README.md
 
-## 🤝 How to Contribute
+## 🤝 How To Contribute
 
-### 1. Contributing to the Software
+### 1. Contributing To The Software
 
 **Types of Contributions:**
 - Bug fixes and issue resolution
@@ -39,27 +27,35 @@ make run_dev  # Runs on port 5009
 2. **Create branch**: `git checkout -b feature/your-feature`
 3. **Make changes** following our coding standards
 4. **Add tests** for new functionality
-5. **Run tests**: `python -m pytest`
+5. **Run tests**: `make shell` then `python -m pytest` (tests must run in Docker)
 6. **Submit PR** with clear description
 
 ### 2. Code Standards
 
+**Keep PRs Small and Focused:**
+- Submit small, focused pull requests for easier review
+- One feature or fix per PR when possible
+- Break large changes into logical, reviewable chunks
+
 Follow Python packaging standards and best practices:
 
 ```python
-def generate_config(prompt: str, model: str = "gpt-4o") -> dict:
-    """Generate CanvasXpress config from natural language.
+def validate_prompt(prompt: str) -> str:
+    """Validate and clean user input prompt.
     
     Args:
-        prompt: Description like "box plot of cty grouped by manufacturer"
-        model: LLM model to use
+        prompt: User's natural language description
         
     Returns:
-        CanvasXpress configuration dictionary
+        Cleaned and validated prompt string
+        
+    Raises:
+        ValueError: If prompt is empty or invalid
     """
-    if not prompt.strip():
+    if not prompt or not prompt.strip():
         raise ValueError("Prompt cannot be empty")
-    return config
+    
+    return prompt.strip()
 ```
 
 **Standards:**
@@ -70,9 +66,12 @@ def generate_config(prompt: str, model: str = "gpt-4o") -> dict:
 
 ## 🧪 Testing Requirements
 
-All contributions must include appropriate tests:
+All contributions must include appropriate tests. **All test commands must be run inside the Docker container:**
 
 ```bash
+# First enter Docker container
+make shell
+
 # Run all tests (uses real APIs if configured, mocks otherwise)
 python -m pytest
 
@@ -155,10 +154,10 @@ For feature requests, please provide:
 ### Getting Help
 
 **Before seeking support:**
-1. Check this README and documentation files
+1. Check [README.md](README.md) and documentation files
 2. Search existing GitHub issues
 3. Review the [API Documentation](docs/API.md) and [Integration Guide](docs/INTEGRATION.md)
[…]
