### 2025-07-21 daf287b
Various small fixes
