### 2025-07-21 ffcda8f
Changed testing to just reference TESTING.md
