### 2025-07-21 854baaf
Added fresh env setup section to explain buildfresh
