### 2025-07-21 7b7fc93
Various small fixes and updates
