### 2025-07-21 c28dce5
Made some moves of text and improvements to the wording
