### 2025-07-21 e9e9f92
Various fixes and enhancements for clarify and flow
