### 2025-07-22 439cc72
Added table of contents with clickable links to sections, and a glossary of LLM terms
