+++ pytest.ini
@@ -14,6 +14,6 @@ addopts =
     --cov-fail-under=80
     --ignore=OLD_BAK
 markers =
-    unit: Unit tests
-    integration: Integration tests
-    slow: Slow running tests
+    unit: marks tests as unit tests
+    integration: marks tests as integration tests
+    slow: marks tests as slow running tests
