### 2025-07-31 #7529 @crvernon
:wave: @xavieryao please give me an update to what is left outstanding for you at this point.
