### 2025-09-05 #7529 @buddyroo30
Hi,
Just wanted to check in about the status of our paper. When might you be able to complete the review?
thanks,Andrew
    On Friday, August 22, 2025 at 02:51:14 PM EDT, Chris Vernon ***@***.***> wrote:

 crvernon left a comment (openjournals/joss-reviews#7529)
👋 @xavieryao just a reminder to please give me an update to what is left outstanding for you at this point.
