### 2025-09-29 #7529 @buddyroo30
Hi,

We have not heard any updates in some time and would appreciate it if you could provide any information on our paper's current status.
Thank you for your time and consideration.

Sincerely,
Andrew
    On Friday, September 5, 2025 at 10:11:57 AM EDT, Andrew Smith ***@***.***> wrote:

  Hi,
Just wanted to check in about the status of our paper. When might you be able to complete the review?
thanks,Andrew
    On Friday, August 22, 2025 at 02:51:14 PM EDT, Chris Vernon ***@***.***> wrote:

 crvernon left a comment (openjournals/joss-reviews#7529)
👋 @xavieryao just a reminder to please give me an update to what is left outstanding for you at this point.
