### 2025-10-01 #7529 @crvernon
👋  @buddyroo30 - I think the submission is in good shape.  However, there are some errors in your BibTex formatting for the paper that I am happy to provide a fix for.  I can 1) post a revised paper.md and paper.bib file for you here; or 2) if you confirm the branch name you are currently making changes on, I can fork your repository and issue the needed changes in the form of a pull request.

After we get this straightened out, I'll provide the final steps for you to address so we can move forward with acceptance.
