### 2025-10-21 #7529 @crvernon
@buddyroo30 after checking this out again and digging into your no date references some, I have decided to move forward with this in it's current state.

So here is what we have left to do:

- [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) if you have not done so already.  Please ensure that the software archive uses the same license as the license you have posted on code repository.

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the people and the order in which they are listed on your paper exactly).

- [ ]  Please respond with the DOI of the archived version here along with the version number released.
