### 2025-10-27 01c0a8f
Fix grammatical error in paper summary: flexible. -> flexible,
