+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Summary
 
-Powerful data visualization packages like CanvasXpress offer rich features for exploring datasets but often have a high learning curve, requiring detailed web development skills. We present a system that combines guided autocomplete and Large Language Models (LLMs) to simplify the use of such packages, focusing specifically on CanvasXpress. Users upload tabular data or start from any CanvasXpress visualization, then describe their desired visualization in plain English—either by following guided suggestions or typing free-form text. Guided autocomplete ensures perfect configuration accuracy, while LLMs serve as a robust backup for flexible. free-form input. The system can also generate large-scale synthetic training data for use as few-shot examples or for fine-tuning. We evaluate accuracy of LLM CanvasXpress configuration generation using thousands of synthetic examples with Retrieval Augmented Generation (RAG), achieving over 97% exact match accuracy. Results show that accuracy increases with more few-shot examples but decreases with prompt complexity, highlighting the challenge of generating complex visualizations from natural language.
+Powerful data visualization packages like CanvasXpress offer rich features for exploring datasets but often have a high learning curve, requiring detailed web development skills. We present a system that combines guided autocomplete and Large Language Models (LLMs) to simplify the use of such packages, focusing specifically on CanvasXpress. Users upload tabular data or start from any CanvasXpress visualization, then describe their desired visualization in plain English—either by following guided suggestions or typing free-form text. Guided autocomplete ensures perfect configuration accuracy, while LLMs serve as a robust backup for flexible, free-form input. The system can also generate large-scale synthetic training data for use as few-shot examples or for fine-tuning. We evaluate accuracy of LLM CanvasXpress configuration generation using thousands of synthetic examples with Retrieval Augmented Generation (RAG), achieving over 97% exact match accuracy. Results show that accuracy increases with more few-shot examples but decreases with prompt complexity, highlighting the challenge of generating complex visualizations from natural language.
 
 # Statement of Need
 
