### 2025-10-28 #7529 @crvernon
@buddyroo30 the title of your paper is:  "Generating Visualizations Conversationally using LLMs" you named your archive "Generating Visualizations Conversationally using Guided Autocomplete and LLMs" - please make them match.
